Suppresion et modif ok par userId

+++ back/controllers/sauces.js
@@ -31,15 +31,21 @@ exports.getOneSauce = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({_id: req.params.id})
   .then(sauce=> {
-    const filename = sauce.imageUrl.split('/images/')[1];
-    fs.unlink(`images/${filename}`,() =>{
-       Sauce.deleteOne({ _id: req.params.id })
-    .then(() => res.status(201).json({ message: "Sauce supprimée !" }))
-    .catch((error) => res.status(500).json({ error }))
-    })
+    if(sauce.userId != req.auth.userId){
+      res.status(401).json({message:' Non autorisé'});
+    } else{
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`,() =>{
+          Sauce.deleteOne({ _id: req.params.id })
+        .then(() => res.status(201).json({ message: "Sauce supprimée !" }))
+        .catch((error) => res.status(500).json({ error }));
+      
+        });
+      }
   })
-  .catch(error => res.status(500).json({ error }));
-};
+  .catch(error => {res.status(500).json({ error });
+  });
+};  
 
 exports.modifySauce = (req, res, next) => {
   const sauceObject = req.file ? {
